requirements.txt
psycopg2-binary>=2.7
h5py>=2.8
matplotlib>=1.4.3
numpy>=1.15.4
pandas>=0.25.1
jinja2>=2.7.3
scipy>=0.15.1
six>=1.9.0
pynrrd>=0.2.1
future >= 0.14.3
requests
requests-toolbelt
simplejson>=3.10.0
scikit-image>=0.14.0
scikit-build
statsmodels==0.9.0
simpleitk
argschema
marshmallow==3.0.0rc6
glymur
xarray
hdmf==1.0.2
pynwb==1.0.2
tables==3.5.1 # pinning this because updates tend to not include wheels immediately
seaborn
